refactored error_msg and error_msg_utils
