+++ src/error_msg.c
@@ -3,95 +3,61 @@
 /*                                                        :::      ::::::::   */
 /*   error_msg.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gyong-si <gyongsi@student.42.fr>           +#+  +:+       +#+        */
+/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/17 14:35:43 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/03 13:24:54 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/10 11:54:29 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-char	*ft_strjoin_free(char **s1, char const *s2)
+static char	*generate_error_message(int error_no)
 {
-	char	*result;
+	char	*error_msg;
 
-	if (!s1 || !s2)
-		return (NULL);
-	if (!*s1)
-	{
-		*s1 = ft_strdup("");
-		if (!*s1)
-			return (NULL);
-	}
-	result = ft_strjoin(*s1, s2);
-	free(*s1);
-	*s1 = NULL;
-	return (result);
+	if (error_no == 42)
+		error_msg = ft_strdup("command not found");
+	else if (error_no == 43)
+		error_msg = ft_strdup("too many arguments");
+	else if (error_no == EISDIR)
+		error_msg = ft_strdup("Is a directory");
+	else if (error_no == ENOENT)
+		error_msg = ft_strdup("No such file or directory");
+	else if (error_no == EACCES)
+		error_msg = ft_strdup("Permission denied");
+	else
+		error_msg = ft_strdup(strerror(error_no));
+	return (error_msg);
 }
 
-void	delete_command(void *elem)
+static int	map_error_to_exit_code(int error_no)
 {
-	t_cmd	*cmd;
+	int	return_no;
 
-	cmd = (t_cmd *)elem;
-	ft_split_free(&cmd->argv);
-	ft_lstclear(&cmd->input, free);
-	ft_lstclear(&cmd->output, free);
-	free(cmd);
+	if (error_no == 42 || error_no == ENOENT)
+		return_no = 127;
+	else if (error_no == EISDIR || error_no == EACCES)
+		return_no = 126;
+	else
+		return_no = 1; // General error
+	return (return_no);
 }
 
-void	free_and_exit(t_shell *minishell, int return_value)
+int	minishell_error_msg(char *cmd, int error_no)
 {
-	rl_clear_history();
-	ft_split_free(&minishell->env);
-	ft_lstclear((t_list **)&(minishell->cmd_list), delete_command);
-	free(minishell->prompt);
-	exit(return_value);
-}
+	char	*error;
+	char	*error_msg;
+	int		return_no;
 
-int minishell_error_msg(char *cmd, int error_no)
-{
-    char *error;
-    int return_no;
-    char *error_msg;
-
-    error = ft_strdup("minishell: ");
-    error = ft_strjoin_free(&error, cmd);
-    error = ft_strjoin_free(&error, ": ");
-
-    // Specific error messages
-    if (error_no == 42)
-        error_msg = ft_strdup("command not found");
-    else if (error_no == 43)
-        error_msg = ft_strdup("too many arguments");
-    else if (error_no == EISDIR) {
-        error_msg = ft_strdup("Is a directory");
-    } else if (error_no == ENOENT) {
-        error_msg = ft_strdup("No such file or directory");
-    } else if (error_no == EACCES) {
-        error_msg = ft_strdup("Permission denied");
-    } else {
-        error_msg = ft_strdup(strerror(error_no));
-    }
-
-    error = ft_strjoin_free(&error, error_msg);
-    free(error_msg);
-    ft_putendl_fd(error, 2);
-    free(error);
-
-    // Map specific errors to their corresponding exit codes
-    if (error_no == 42) {  // "command not found"
-        return_no = 127;
-    } else if (error_no == ENOENT) {  // "No such file or directory"
-        return_no = 127;
-    } else if (error_no == EISDIR) {  // "Is a directory"
-        return_no = 126;
-    } else if (error_no == EACCES) {  // "Permission denied"
-        return_no = 126;
-    } else {
-        return_no = 1;  // General error
-    }
-
-    return return_no;
+	error = ft_strdup("minishell: ");
+	error = ft_strjoin_free(&error, cmd);
+	error = ft_strjoin_free(&error, ": ");
+	error_msg = generate_error_message(error_no);
+	error = ft_strjoin_free(&error, error_msg);
+	free(error_msg);
+	ft_putendl_fd(error, 2);
+	free(error);
+	return_no = map_error_to_exit_code(error_no);
+	return (return_no);
 }
