+++ src/pipex/pipex_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   pipex_utils.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gyong-si <gyongsi@student.42.fr>           +#+  +:+       +#+        */
+/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 10:23:30 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/06/03 13:10:36 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/10 12:02:46 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
